chore(script folder)
- 스크리비트 폴더 정리 및 게임 씬 UI 컨트롤러 추가
