+++ Assets/Script/GameScene.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cc7604d7c8946aa47923a1896f0bcafd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
