+++ Assets/Script/GameScene/GameSceneUIController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 08aa0394380fbc840b10175adb4f8445
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
