+++ Assets/Script/LobbyScene.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6ef00d7d39392524eb7c226fecb14b50
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
