+++ Assets/Script/MatchedInfoDto.cs
@@ -1,0 +1,8 @@
+[System.Serializable]
+public class MatchedInfoDto
+{
+    public string message;
+    public string user1;
+    public string user2;
+    public string sessionId;
+}
