+++ Assets/Script/MatchedInfoDto.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d43d6163c1add3348ad05695af2c53e6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
